S-40: Restructured folders as described.
