+++ closetr-web/src/app/app-routing.module.ts
@@ -1,17 +1,17 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LoginComponent } from './login/login.component';
-import { RegisterComponent } from './register/register.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { ClosetManageComponent } from './closet-manage/closet-manage.component';
-import { AddClothingComponent } from './add-clothing/add-clothing.component';
-import { EditClothingComponent } from './edit-clothing/edit-clothing.component';
-import { LogOutfitComponent } from './today-widget/log-outfit/log-outfit.component';
-import { SpendingManageComponent } from './spending-manage/spending-manage.component';
-import { BudgetManageComponent } from './budget-manage/budget-manage.component';
-import { ProfileComponent } from './profile/profile.component';
+import { LoginComponent } from './pages/page-full/login/login.component';
+import { RegisterComponent } from './pages/page-full/register/register.component';
+import { DashboardComponent } from './pages/page-full/dashboard/dashboard.component';
+import { ClosetManageComponent } from './pages/page-full/closet-manage/closet-manage.component';
+import { AddClothingComponent } from './pages/page-full/add-clothing/add-clothing.component';
+import { EditClothingComponent } from './pages/page-full/edit-clothing/edit-clothing.component';
+import { LogOutfitComponent } from './pages/page-full/log-outfit/log-outfit.component';
+import { SpendingManageComponent } from './pages/page-full/spending-manage/spending-manage.component';
+import { BudgetManageComponent } from './pages/page-full/budget-manage/budget-manage.component';
+import { ProfileComponent } from './pages/page-full/profile/profile.component';
 import { AppComponent } from './app.component';
-import { AuthGuard } from './_guards/auth.guard';
+import { AuthGuard } from './guards/auth.guard';
 
 const routes: Routes = [
   {path: '', canActivate:[AuthGuard], children: [
