+++ closetr-web/src/app/pages/page-full/add-clothing/add-clothing.component.html
