+++ closetr-web/src/app/pages/page-full/base-general/base-general.component.ts
