+++ closetr-web/src/app/pages/page-full/edit-clothing/edit-clothing.component.spec.ts
