+++ closetr-web/src/app/pages/page-full/profile/profile.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthenticationService } from '../services/authentication.service';
-import { UserService } from '../services/user.service';
-import { User } from '../models/user.model';
+import { AuthenticationService } from '../../../services/authentication.service';
+import { UserService } from '../../../services/user.service';
+import { User } from '../../../models/user.model';
 import { Subscription } from 'rxjs';
 
 @Component({
