+++ closetr-web/src/app/pages/page-full/register/register.component.html
