+++ closetr-web/src/app/pages/page-full/register/register.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { RouterModule, Routes, Router } from '@angular/router';
-import { UserService } from '../services/user.service';
-import { User } from '../models/user.model';
-import { AuthenticationService } from '../services/authentication.service';
+import { UserService } from '../../../services/user.service';
+import { User } from '../../../models/user.model';
+import { AuthenticationService } from '../../../services/authentication.service';
 
 @Component({
   selector: 'app-register',
