+++ closetr-web/src/app/pages/page-full/spending-manage/spending-manage.component.spec.ts
