+++ closetr-web/src/app/pages/page-partial/closet-card/closet-card.component.html
