+++ closetr-web/src/app/pages/page-partial/closet-widget/closet-widget.component.html
