+++ closetr-web/src/app/pages/page-partial/closet-widget/closet-widget.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { ClosetService } from '../services/closet.service';
-import { AuthenticationService } from '../services/authentication.service';
-import { Clothing } from '../models/clothing.model';
-import { User } from '../models/user.model';
+import { ClosetService } from '../../../services/closet.service';
+import { AuthenticationService } from '../../../services/authentication.service';
+import { Clothing } from '../../../models/clothing.model';
+import { User } from '../../../models/user.model';
 import { Subscription } from 'rxjs';
 
 @Component({
